move back was no longer working
git-svn-id: https://svn.code.sf.net/p/zvtm/code/zgrviewer/trunk@4407 74e165f3-c697-40a6-91af-552c8936d7fe

+++ src/main/java/net/claribole/zgrviewer/GraphicsManager.java
@@ -523,7 +523,7 @@ public class GraphicsManager implements ComponentListener, CameraListener, Java2
             Animation at = animator.getAnimationFactory().createCameraTranslation(ConfigManager.ANIM_MOVE_LENGTH, mSpace.getCamera(0),
                 (Point2D.Double)animParams.elementAt(1), true, SlowInSlowOutInterpolator.getInstance(), null);
             Animation aa = animator.getAnimationFactory().createCameraAltAnim(ConfigManager.ANIM_MOVE_LENGTH, mSpace.getCamera(0),
-                (Float)animParams.elementAt(0), true, SlowInSlowOutInterpolator.getInstance(), null);
+                (Double)animParams.elementAt(0), true, SlowInSlowOutInterpolator.getInstance(), null);
             animator.startAnimation(at, false);
             animator.startAnimation(aa, false);
             previousLocations.removeElementAt(previousLocations.size()-1);
